Move common test stuff into separate files
